+++ upload/admin/language/english/module/pvnm_testimonials_widget.php
@@ -1,0 +1,33 @@
+<?php
+// Heading
+$_['heading_title']					= 'Perfect Testimonials Widget';
+
+// Tab
+$_['tab_settings']					= 'Settings';
+$_['tab_help']						= 'Help';
+
+// Text
+$_['text_module']					= 'Modules';
+$_['text_success']					= 'Success: You have modified module!';
+$_['text_edit']						= 'Edit Module';
+$_['text_documentation']			= 'Documentation';
+$_['text_developer']				= 'Developer';
+$_['text_tab_module']				= 'Module';
+
+// Button
+$_['button_testimonials']			= 'Testimonials';
+$_['button_votes']					= 'Votes';
+$_['button_settings']				= 'Settings';
+$_['button_widget']					= 'Widget';
+$_['button_add']					= 'Add Module';
+
+// Entry
+$_['entry_name']					= 'Module Name';
+$_['entry_rating_status']			= 'Block Rating Status';
+$_['entry_testimonials_status']		= 'Block Testimonials Status';
+$_['entry_testimonials_title']		= 'Block Testimonials Title';
+$_['entry_testimonials_limit']		= 'Limit Testimonials';
+$_['entry_status']					= 'Module Status';
+
+// Error
+$_['error_permission']				= 'Warning: You do not have permission to modify module!';