+++ upload/admin/language/russian/module/pvnm_testimonials_widget.php
@@ -1,0 +1,33 @@
+<?php
+// Heading
+$_['heading_title']					= 'Perfect Testimonials Widget';
+
+// Tab
+$_['tab_settings']					= 'Настройки';
+$_['tab_help']						= 'Помощь';
+
+// Text
+$_['text_module']					= 'Модули';
+$_['text_success']					= 'Настройки модуля обновлены!';
+$_['text_edit']						= 'Настройки модуля';
+$_['text_documentation']			= 'Документация';
+$_['text_developer']				= 'Разработчик';
+$_['text_tab_module']				= 'Модуль';
+
+// Button
+$_['button_testimonials']			= 'Отзывы';
+$_['button_votes']					= 'Голоса';
+$_['button_settings']				= 'Настройки';
+$_['button_widget']					= 'Виджет';
+$_['button_add']					= 'Добавить модуль';
+
+// Entry
+$_['entry_name']					= 'Название модуля';
+$_['entry_rating_status']			= 'Блок рейтинга';
+$_['entry_testimonials_status']		= 'Блок отзывов';
+$_['entry_testimonials_title']		= 'Заголовок блока отзывов';
+$_['entry_testimonials_limit']		= 'Лимит отзывов';
+$_['entry_status']					= 'Статус модуля';
+
+// Error
+$_['error_permission']				= 'У вас нет прав для изменения этого модуля!';