+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+#### 1.1.0
+* Added Widget Latest Testimonials and Store Rating
+* In admin menu added link to Perfect Testimonials Widget
+
 #### 1.0.0
 * Initial release